Crate cfg_types

Common-types of the Centrifuge chain.

Modules

adjustments

consts

domain_address

epoch

fee_keys

fixed_point

Decimal Fixed Point implementations for Substrate runtime. Copied over from sp_arithmetic

ids

A module that contains all ids where we should be REALLY careful when changing them.

investments

locations

oracles

orders

permissions

pools

time

tokens

xcm

Type Aliases

EVMChainId

The EVM Chain ID The type should accommodate all chain ids listed on https://chainlist.org/.

ParaId

A raw para ID