Module cfg_types::permissions

Structs

CurrencyAdminRoles

The current admin roles we support

PermissionRoles

The structure that we store in the pallet-permissions storage This here implements trait Properties.

PermissionedCurrencyHolderInfo

PermissionedCurrencyHolders

PoolAdminRoles

The current admin roles we support

TrancheInvestorInfo

TrancheInvestors

Enums

PermissionScope

PermissionedCurrencyRole

PoolRole

PoolRole can hold any type of role specific functions a user can do on a given pool.

Role

The Role enum is used by the permissions pallet, to specify which role an account has within a specific scope.

Constants

UNION

The implementation of trait Properties for our PermissionsRoles does not care which Moment is passed to the PoolRole::TrancheInvestor(TrancheId, Moment) variant. This UNION shall reflect that and explain to the reader why it is passed here.